requirements.txt
selenium==3.*
requests==2.*
percy-python-selenium==0.*
